Add useEffect hook to App.js
