+++ src/Components/RocketList/RocketList.js
@@ -1,6 +1,5 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { fetchRockets } from '../../redux/configureStore';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import Rocket from '../Rocket/Rocket';
 // import styles
 import styles from './RocketList.module.css';
@@ -8,12 +7,6 @@ import styles from './RocketList.module.css';
 const RocketList = () => {
   // Get rockets from the store
   const rockets = useSelector((state) => state.rockets);
-  // Get dispatch function
-  const dispatch = useDispatch();
-  // Get rockets from the API
-  useEffect(() => {
-    dispatch(fetchRockets());
-  }, [dispatch]);
 
   return (
     <div className={styles['rocket-list-container']}>
