+++ src/redux/configureStore.js
@@ -1,7 +1,9 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
+import logger from 'redux-logger';
 import rocketsReducer, { fetchRockets } from './rockets/rockets';
 import dragonsReducer, { fetchDragons } from './dragons/dragons';
 import missionsReducer, { fetchMissions } from './missions/missions';
+// Logger with default options
 
 const rootReducer = combineReducers({
   rockets: rocketsReducer,
@@ -11,6 +13,8 @@ const rootReducer = combineReducers({
 
 const store = configureStore({
   reducer: rootReducer,
+  // middleware: applyMiddleware(logger),
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 });
 
 export { fetchRockets, fetchDragons, fetchMissions };
